requirements.txt
streamlit
torch
opencv-python
Pillow
numpy
matplotlib
pyyaml
requests
